remove Utils class
